requirement.txt
streamlit==1.26.0
Pillow==9.4.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.1
tensorflow==2.12.0
keras==2.12.0
